refactor(planner): split engine.py into engine_core, engine_explanation, engine_hardware
